add app + side bar to all bottom nav pages
idk if this is a good idea? might change back once we start actually working on ui
