+++ mobile/lib/mainpage.dart
@@ -1,5 +1,6 @@
 import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:trill/pages/home.dart';
@@ -94,20 +95,136 @@ class _MainPageState extends State<MainPage> {
     return !_userInfoSet
         ? Scaffold(body: LoadingScreen())
         : Scaffold(
-            // temp appbar for button to test signing out
-            // todo: implement button actually
-            /*appBar: AppBar(
-              actions: [
-                MaterialButton(
-                  onPressed: () {
-                    Amplify.Auth.signOut().then((_) {
-                      Navigator.pushReplacementNamed(context, '/login');
-                    });
-                  },
-                  child: Icon(Icons.logout),
+            appBar: AppBar(),
+            drawer: Drawer(
+              child: Container(
+                color: Color(0xFF1A1B29),
+                child: ListView(
+                  padding: EdgeInsets.zero,
+                  children: [
+                    SizedBox(height: 30),
+                    Container(
+                      padding: EdgeInsets.fromLTRB(100, 0, 100, 0),
+                      child: ClipRRect(
+                          borderRadius: BorderRadius.circular(1000.0),
+                          child: Image.asset("images/gerber.jpg",
+                              fit: BoxFit.cover)),
+                    ),
+                    SizedBox(height: 10),
+                    Column(children: [
+                      const Text(
+                        'Matthew Gerber',
+                        style: TextStyle(
+                          color: Colors.blue,
+                          fontSize: 20,
+                          fontWeight: FontWeight.bold,
+                        ),
+                      ),
+                      const Text('@GerbersGrumblings'),
+                    ]),
+                    SizedBox(height: 20),
+                    Row(
+                      children: [
+                        Spacer(),
+                        Container(
+                          padding: EdgeInsets.all(5),
+                          decoration: BoxDecoration(
+                            border:
+                                Border.all(color: Color(0xFFBC6AAB), width: 1),
+                            borderRadius: BorderRadius.all(Radius.circular(30)),
+                          ),
+                          child: Center(
+                            child: RichText(
+                              text: TextSpan(
+                                  text: 'Followers: 100',
+                                  style: TextStyle(
+                                      fontSize: 11,
+                                      fontWeight: FontWeight.bold),
+                                  recognizer: TapGestureRecognizer()
+                                    ..onTap = () => Navigator.pushNamed(
+                                        context, '/followers')),
+                            ),
+                          ),
+                        ),
+                        Spacer(flex: 2),
+                        Container(
+                          padding: EdgeInsets.all(5),
+                          decoration: BoxDecoration(
+                            border:
+                                Border.all(color: Color(0xFFBC6AAB), width: 1),
+                            borderRadius: BorderRadius.all(Radius.circular(30)),
+                          ),
+                          child: Center(
+                            child: RichText(
+                              text: TextSpan(
+                                  text: 'Following: 100',
+                                  style: TextStyle(
+                                      fontSize: 11,
+                                      fontWeight: FontWeight.bold),
+                                  recognizer: TapGestureRecognizer()
+                                    ..onTap = () => Navigator.pushNamed(
+                                        context, '/following')),
+                            ),
+                          ),
+                        ),
+                        Spacer()
+                      ],
+                    ),
+                    SizedBox(height: 20),
+                    ListTile(
+                      leading: Icon(Icons.home_outlined, color: Colors.white),
+                      title: const Text('Home'),
+                      onTap: () {
+                        Navigator.pushNamed(context, '/main');
+                      },
+                    ),
+                    ListTile(
+                      leading: Icon(Icons.library_music_outlined,
+                          color: Colors.white),
+                      title: const Text('Albums'),
+                      onTap: () {
+                        Navigator.pushNamed(context, '/albums');
+                      },
+                    ),
+                    ListTile(
+                      leading:
+                          Icon(Icons.rate_review_outlined, color: Colors.white),
+                      title: const Text('Reviews'),
+                      onTap: () {
+                        // TODO: Change back to /reviews
+                        Navigator.pushNamed(context, '/review');
+                      },
+                    ),
+                    ListTile(
+                      leading: Icon(Icons.format_list_bulleted_outlined,
+                          color: Colors.white),
+                      title: const Text('Listen Later'),
+                      onTap: () {
+                        Navigator.pushNamed(context, '/listenlater');
+                      },
+                    ),
+                    ListTile(
+                      leading: Icon(Icons.favorite_outline_outlined,
+                          color: Colors.white),
+                      title: const Text('Likes'),
+                      onTap: () {
+                        Navigator.pushNamed(context, '/likes');
+                      },
+                    ),
+                    SizedBox(height: 30),
+                    ListTile(
+                        leading:
+                            Icon(Icons.logout_outlined, color: Colors.white),
+                        title: const Text('Log Out'),
+                        onTap: () {
+                          Amplify.Auth.signOut().then((_) {
+                            Navigator.pushReplacementNamed(context, '/login');
+                          });
+                        }),
+                  ],
                 ),
-              ],
-            ),*/
+              ),
+            ),
             // IndexedStack keeps the states of each page
             body: IndexedStack(
               index: currentIndex,
